Added 'Festo Flexible' as a finger type
Added 'Festo Flexible' to the finger comboxbox and to the required dictionaries.
Put "TODO" comments where real values need to be inserted later on for the 'Festo Flexible' dictionary values (they are currently not known).
ControlScheme.py: real gain values need to be determined and inserted into the update_finger_gains() method at a later date.