Forced user to manually select a finger type
When the program is first opened, no fingertype option is selected in the finger combobox. 
A image of the gripper with no fingers on it is shown.
If the user attempts to hit open or close gripper, an error message is shown that says please select a finger type (and the gripper will refuse to move until one is selected). This is done by checks to see if the fingertype is none at the start of the close_button_clicked() and open_button_clicked() methods.
Put str() around self.kukaGripper1.fingertype in the update_gripper_finger_image() method of GripperGUI.py so that when fingertype is None, it can access the image associated with None in the image dictionary