Handed exception when image file is not found
The update_gripper_finger_image() method in GripperGUI.py opens an image of the gripper fingers. A FileNotFound exception was thrown if the image was not in the expected location/the image path name was wrong when trying to open the image. 
Put in a try except block to catch and handle this exception. If the file cannot be found, the image currently shown by the label is removed (set to ""), and is replaced by text that says "Image of gripper fingers could not be found".
The method exits after that.
Changed the background color of the label so it matches.