Improved motor error
Before in Gripper.py, the status never could update to 'Error' because it was accidentally written by the senior team as self.status == 'Error'. Fixed that
Added a motor error dictionary so I can give a more specific error message on the GUI. Unsure if I got the right values from the datasheet so left it as it is currently.
The motor error code is now shown on the GUI.