Made gripper status an enum
Made the gripper status variable into an enumerated type to prevent catastrophic errors. 
I had found mistypings/uniform capitalization in the original code that caused the comparision checks to fail. This will ensure that only controlled status options are introduced, especially if the code is modified further in the future.